+++ src/data/__init__.py
@@ -1,0 +1,2 @@
+
+ALLOWED_BASE_DATASETS = ["mini_imagenet", "omniglot", "flowers102", "stanford_cars"]